chore: refactored handleCommandString to avoid circular dependency
